[#25] chore : AccountInfo Composable 함수들을 ui/component 패키지로 이동
